implement Global exception handler + alert clean up

+++ src/main/java/com/blog/myblog/dto/appUser/UserEditDTO.java
@@ -9,6 +9,6 @@ public class UserEditDTO {
 
     private String firstName;
     private String lastName;
-    private Short version;
+    private int version;
 
 }
