+++ src/main/java/com/blog/myblog/dto/article/ArticleEditDTO.java
@@ -9,5 +9,5 @@ public class ArticleEditDTO {
 
     private String title;
     private String body;
-    private Short version;
+    private int version;
 }
